Move check_mod_attr query in librustc_passes.
