+++ src/librustc/hir/mod.rs
@@ -10,7 +10,6 @@ pub mod upvars;
 use crate::ty::query::Providers;
 
 pub fn provide(providers: &mut Providers<'_>) {
-    check_attr::provide(providers);
     map::provide(providers);
     upvars::provide(providers);
 }
